Bugfix circular import.
The startup file was importing controller modules, and controller modules will have a reference to the container to get dependencies, this causes a circular import between modules. To solve it, every related functionality to the API is being moved to the main module.
