+++ startup.py
@@ -1,11 +1,7 @@
 import os
-from fastapi import FastAPI
-from fastapi.middleware.cors import CORSMiddleware
 from dependency_injector import containers, providers
 
 from data.database import Database
-from features.tickets import tickets_controller
-from features.messages import messages_controller
 
 class DependencyContainer(containers.DeclarativeContainer):
     wiring_config = containers.WiringConfiguration(modules=[])
@@ -33,17 +29,3 @@ def add_configuration_providers(container: DependencyContainer):
     container.configuration.env.from_value(APP_ENVIRONMENT)
     container.configuration.connection_string.from_env('APP_DB_CONNECTION')
 
-def configure_pipeline(app: FastAPI) -> FastAPI:
-    # Middlewares
-    app.add_middleware(
-        CORSMiddleware,
-        allow_origins=["*"],
-        allow_methods=["*"],
-        allow_headers=["*"],
-    )
-
-    # Routes
-    app.include_router(tickets_controller.router)
-    app.include_router(messages_controller.router)
-
-    return app
